isban check on educationController

+++ controllers/user/educationController.js
@@ -3,10 +3,18 @@ const Educations = require("../../models/educations");
 const User = require("../../models/user");
 
 const addEducation = async (req, res) => {
-  const { userId } = req.params;
   const { institution, course, startYear, endYear } = req.body;
 
   try {
+    const userId = req.user.id;
+    const user = await User.findOne({
+      where: { id: userId },
+      attributes: ["isBanned"],
+    });
+
+    if (user.isBanned) {
+      return res.status(403).json({ error: "User account is banned" });
+    }
     const newEducation = await Educations.create({
       userId,
       institution,
@@ -54,7 +62,16 @@ const updateEducation = async (req, res) => {
   const { institution, course, startYear, endYear } = req.body;
 
   try {
-    const education = await Educations.findOne({ where: { id } });
+    const userId = req.user.id;
+    const user = await User.findOne({
+      where: { id: userId },
+      attributes: ["isBanned"],
+    });
+
+    if (user.isBanned) {
+      return res.status(403).json({ error: "User account is banned" });
+    }
+    const education = await Educations.findOne({ where: { id, userId } });
 
     if (!education) {
       return res.status(404).json({ error: "education not found" });
@@ -77,7 +94,23 @@ const updateEducation = async (req, res) => {
 const deleteEducation = async (req, res) => {
   const { id } = req.params;
   try {
-    await Educations.destroy({ where: { id } });
+    const userId = req.user.id;
+    const user = await User.findOne({
+      where: { id: userId },
+      attributes: ["isBanned"],
+    });
+
+    if (user.isBanned) {
+      return res.status(403).json({ error: "User account is banned" });
+    }
+    const education = await Educations.findOne({
+      where: { id, userId },
+    });
+
+    if (!education) {
+      return res.status(404).json({ error: "Education not found" });
+    }
+    await Educations.destroy({ where: { id, userId } });
     res.status(200).json({ message: "Education deleted successfully" });
   } catch (error) {
     console.error("Error deleting education", error);
